feat(server-utils)!: Rename `otlpIntegration` to `openTelemetryIntegration`
