+++ dev-packages/e2e-tests/test-applications/nextjs-otlp/sentry.server.config.ts
@@ -12,5 +12,5 @@ Sentry.init({
   enableOpenTelemetrySetup: false,
 
   // Puts the active OpenTelemetry span's trace on everything Sentry sends.
-  integrations: [Sentry.otlpIntegration()],
+  integrations: [Sentry.openTelemetryIntegration()],
 });
