+++ packages/core/src/carrier.ts
@@ -43,7 +43,7 @@ export interface SentryCarrier {
   /** Strategy for assembling segment spans into transactions; set by SDKs that defer capture. */
   segmentSpanCaptureStrategy?: SegmentSpanCaptureStrategy;
 
-  /** Supplies trace context from a non-Sentry source (e.g. OpenTelemetry); set by `otlpIntegration`. */
+  /** Supplies trace context from a non-Sentry source (e.g. OpenTelemetry); set by `openTelemetryIntegration`. */
   externalPropagationContextProvider?: () => { traceId: string; spanId: string } | undefined;
 
   /** Overwrites TextEncoder used in `@sentry/core`, need for `react-native@0.73` and older */
