+++ packages/google-cloud-serverless/src/index.ts
@@ -118,7 +118,7 @@ export {
   postgresIntegration,
   postgresJsIntegration,
   prismaIntegration,
-  otlpIntegration,
+  openTelemetryIntegration,
   getOtlpTracesEndpoint,
   processSessionIntegration,
   hapiIntegration,
