+++ packages/server-utils/src/exports.ts
@@ -1,5 +1,5 @@
 // Shared exports not using diagnostics channels
 export { setHttpServerSpanRouteAttribute } from './utils/setHttpServerSpanRouteAttribute';
 export { setAsyncLocalStorageAsyncContextStrategy } from './async-context';
-export { otlpIntegration, getOtlpTracesEndpoint } from './otlp';
+export { openTelemetryIntegration, getOtlpTracesEndpoint } from './opentelemetry';
 export * from './ai';
